Add the pintos-test script
pintos-test is a utility script introduced during Spring 2021
that makes it much easier for students to run and debug individual
test cases, saving them from having to manually copy/paste the
commands to do this from the output of 'make check` like they had to
in the dark-ages. As it was new for Spring 2021 but was not completed
until a couple weeks into the semester, it was placed in the group0
repository and instructions were included in project-0 to have students
copy it from there to the '~/.bin/' directory. The script has been moved
here so that it will automatically be present on student VMs for all
future semesters. It's important to note that in order for pintos-test
to work correctly, some tweaks were made to the pintos Makefiles (see
the git history of the pintos-anon repository
[https://github.com/Berkeley-CS162/pintos-anon] for more details). I've
done my best to propagate these changes to all of the relevant places
(staff solutions, skeletons, etc.) but if you ever find that pintos-test
isn't working for some reason, check that the these changes have been
applied.